001-Reverse_IP_lookup_failure_leads_to_fatal_error by Eric Lammerts <eric@lammerts.org>
DP: When trying out msntp with an IP number as argument, msntp does a
DP: gethostbyaddr() on that, and if it fails, it says "unable to locate IP
DP: address/number" and exits. There does not seem to be any reason to do
DP: a reverse lookup (other than printing the hostname to stderr).